feat(): create token jwt provider
